rename pkg to main
