+++ src/integrationtest/assert_type/HardAssert_test.go
@@ -1,9 +1,9 @@
 package assert_type
 
 import (
-	"github.com/sku0x20/assertG/src/pkg/assert_type"
-	"github.com/sku0x20/assertG/src/pkg/message"
-	"github.com/sku0x20/assertG/src/pkg/message/verbs"
+	"github.com/sku0x20/assertG/src/main/assert_type"
+	"github.com/sku0x20/assertG/src/main/message"
+	"github.com/sku0x20/assertG/src/main/message/verbs"
 	"github.com/sku0x20/gRunner/src/pkg/runner"
 	"strings"
 	"testing"
