+++ src/main/assert/assert.go
@@ -1,6 +1,6 @@
 package assert
 
-import "github.com/sku0x20/assertG/src/pkg/assert_type"
+import "github.com/sku0x20/assertG/src/main/assert_type"
 
 func NewAssert(a assert_type.AssertType) *Assert {
 	return &Assert{a: a}
