+++ src/main/assert/string.go
@@ -1,8 +1,8 @@
 package assert
 
 import (
-	"github.com/sku0x20/assertG/src/pkg/assert_type"
-	"github.com/sku0x20/assertG/src/pkg/asserter"
+	"github.com/sku0x20/assertG/src/main/assert_type"
+	"github.com/sku0x20/assertG/src/main/asserter"
 )
 
 func (a *Assert) ThatString(value string) *asserter.String {
