+++ src/main/assert_type/AssertType.go
@@ -1,6 +1,6 @@
 package assert_type
 
-import "github.com/sku0x20/assertG/src/pkg/message"
+import "github.com/sku0x20/assertG/src/main/message"
 
 type AssertType interface {
 	FailWith(msg *message.Message)
