+++ src/main/assert_type/SoftAssert.go
@@ -1,7 +1,7 @@
 package assert_type
 
 import (
-	"github.com/sku0x20/assertG/src/pkg/message"
+	"github.com/sku0x20/assertG/src/main/message"
 	"testing"
 )
