+++ src/main/equator/ComparableEquator.go
