+++ src/main/equator/Equator.go
