+++ src/test/assert/any_test.go
@@ -1,9 +1,9 @@
 package assert
 
 import (
-	assertP "github.com/sku0x20/assertG/src/pkg/assert"
-	"github.com/sku0x20/assertG/src/pkg/assert_type"
-	"github.com/sku0x20/assertG/src/pkg/asserter"
+	assertP "github.com/sku0x20/assertG/src/main/assert"
+	"github.com/sku0x20/assertG/src/main/assert_type"
+	"github.com/sku0x20/assertG/src/main/asserter"
 	"github.com/sku0x20/gRunner/src/pkg/runner"
 	"testing"
 )
