+++ src/test/assert/assert_test.go
@@ -1,8 +1,8 @@
 package assert
 
 import (
-	assertP "github.com/sku0x20/assertG/src/pkg/assert"
-	"github.com/sku0x20/assertG/src/pkg/assert_type"
+	assertP "github.com/sku0x20/assertG/src/main/assert"
+	"github.com/sku0x20/assertG/src/main/assert_type"
 	"testing"
 )
