+++ src/test/asserter/String_test.go
@@ -1,8 +1,8 @@
 package asserter
 
 import (
-	"github.com/sku0x20/assertG/src/pkg/assert_type"
-	"github.com/sku0x20/assertG/src/pkg/asserter"
+	"github.com/sku0x20/assertG/src/main/assert_type"
+	"github.com/sku0x20/assertG/src/main/asserter"
 	"github.com/sku0x20/gRunner/src/pkg/runner"
 	"testing"
 )
