+++ src/test/message/Message_test.go
@@ -1,7 +1,7 @@
 package message
 
 import (
-	"github.com/sku0x20/assertG/src/pkg/message"
+	"github.com/sku0x20/assertG/src/main/message"
 	"github.com/sku0x20/gRunner/src/pkg/runner"
 	"testing"
 )
